chore: relocate config to project root
This commit moves the config from internal to root for public access
